fix(windows): a fault the pass did not anticipate must not take the queue down
Three narrow holes in the "one damaged package fails alone" guarantee, each of
which would have stopped a whole delivery pass:

A package another process holds open is now retryable rather than an escaping
IOException. Unreadable right now is not the same as wrong, and only the second
deserves to be terminal.

A barrier that cannot be issued now fails the enqueue instead of being shrugged
off, because recording a delivery for bytes that may not have reached the disk
is exactly the state this queue exists to make impossible.

A drain isolates every fault per delivery, not only the ones the queue knows how
to name. Stopping at the first unexpected exception loses every delivery behind
it, which is far worse than one archive waiting for the next pass.

+++ windows/Sources/JazzCaptureCore/Delivery/ArchiveDeliveryRecord.cs
@@ -258,13 +258,13 @@ public sealed record ArchiveDeliveryRecord
 
         var value = new JsonObject
         {
-            [SchemaVersionKey] = SchemaVersion,
+            [SchemaVersionKey] = (long)SchemaVersion,
             [DeliveryIdKey] = DeliveryId,
             [ArchiveIdKey] = ArchiveId,
             [OriginIdKey] = OriginId,
             [CaptureIdsKey] = captures,
-            [FormatVersionKey] = FormatVersion,
-            [RevisionKey] = Revision,
+            [FormatVersionKey] = (long)FormatVersion,
+            [RevisionKey] = (long)Revision,
             [ContentDigestKey] = ContentDigest,
             [RawSha256Key] = RawSha256,
             [ByteLengthKey] = ByteLength,
@@ -272,7 +272,7 @@ public sealed record ArchiveDeliveryRecord
             [TransportKey] = Transport,
             [MappingVersionKey] = MappingVersion,
             [StateKey] = DeliveryStates.ToWire(State),
-            [AttemptKey] = Attempt,
+            [AttemptKey] = (long)Attempt,
             [QueuedAtKey] = QueuedAt,
             [UpdatedAtKey] = UpdatedAt,
         };
